# shared/openapi.yaml
openapi: 3.1.0
info:
  title: cinghialapp API
  version: 0.1.0
  description: |
    D&D 5e campaign manager API.
    Server-authoritative dice. Role-based visibility (player/master).
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /campaigns/{id}/party:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:   { tags: [group], responses: { '200': { description: ok } } }
    patch: { tags: [group], responses: { '200': { description: ok } } }
  /campaigns/{id}/loot: { parameters: [{ $ref: '#/components/parameters/Id' }], get: { tags: [group], responses: { '200': { description: list } } }, post: { tags: [group], responses: { '201': { description: created } } } }
  /campaigns/{id}/quests: { parameters: [{ $ref: '#/components/parameters/Id' }], get: { tags: [group], responses: { '200': { description: list } } }, post: { tags: [group], responses: { '201': { description: created } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...